+++ src/cmdline.py
@@ -1,34 +1,12 @@
 """
-**********************************************************************
-cmdline.py
-
 Copyright (C) 2011-2016 Casper Steinmann
-
-This file is part of the FragIt project.
-
-FragIt is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
-(at your option) any later version.
-
-FragIt is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software
-Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-02110-1301, USA.
-***********************************************************************/
 """
-# python 2.6
 import sys
-from optparse import OptionParser, OptionGroup
+import argparse
 
 from util import fileToMol, file_basename
 from fragmentation import Fragmentation
-from outputformats import get_writer_and_extension
+from outputformats import get_writer_and_extension, supported_output_formats
 
 import strings
 
@@ -40,92 +18,98 @@ def main(argv=None):
     from config import FragItConfig
     cfg = FragItConfig()
 
-    parser = OptionParser(usage=strings.usage,
-                description=strings.description,
-                version=strings.version_str)
-    parser.add_option("-o", "--output", dest="outputfile", type=str, default="", metavar="filename")
+    parser = argparse.ArgumentParser(version=strings.version_str,
+                                     description=strings.__doc__)
+                                     #usage=strings.usage)
 
-    configuration = OptionGroup(parser, "Configuration")
-    general = OptionGroup(parser, "Fragmentation")
-    output = OptionGroup(parser, "Output")
+    parser.add_argument("-o", "--output", dest="outputfile", type=str, default="", metavar="FILENAME")
+    parser.add_argument("input", type=str, metavar="INPUTFILE")
 
-    configuration.add_option("--use-config", dest="useconfigfile", type=str, default="", metavar="filename",
+
+    configuration_group = parser.add_argument_group('Configuration File Options', description="""
+It is possible to use configuration files with FragIt.
+This removes some issues with having many options defined on the commandline.
+To use FragIt as an API all options must also be provided in the form of a configuration file.
+The simplest way to use configuration files is to first generate it using the fragit-conf program.
+Then a configuration file is use by specifying the --use-config flag when running fragit.""")
+    configuration_group.add_argument("--use-config", dest="useconfigfile", type=str, default="", metavar="FILENAME",
                     help="Specify configuration file to use. This will ignore other command line parameters.")
-    configuration.add_option("--make-config", dest="makeconfigfile", type=str, default="", metavar="filename",
-                    help="Specify a filename to use as a configuration file. Use command line options to modify defaults. It is possible to use this command without specifying an input file to generate a clean configuration file.")
 
-    general.add_option("-m", "--maxfragsize", dest="maxFragmentSize", type=int, default=cfg.getMaximumFragmentSize(),metavar="integer",
-                help="The maximum fragment size allowed [default: %default]")
-    general.add_option("-g", "--groupcount", dest="groupcount", type=int, default=cfg.getFragmentGroupCount(),metavar="integer",
-                help="Specify number of consecutive fragments to combine into a single fragment [default: %default]")
-    general.add_option("--disable-protection", dest="disable_protection", action="store_true", default=False,
+    general_group = parser.add_argument_group("Fragmentation Options", description="""
+Controls how to fragmentat the provided input molecule.
+From the command line the default fragmentation patterns will be used.
+It is recommended to use configuration file options as outlined above
+for increased flexibility.
+""")
+    general_group.add_argument("-m", "--maxfragsize", dest="maxFragmentSize", type=int, default=cfg.getMaximumFragmentSize(),metavar="MAX FRAGMENT SIZE",
+                help="The maximum fragment size allowed [default: %(default)s]")
+    general_group.add_argument("-g", "--groupcount", dest="groupcount", type=int, default=cfg.getFragmentGroupCount(),metavar="GROUP n FRAGMENTS INTO ONE",
+                help="Specify number of consecutive fragments to combine into a single fragment [default: %(default)s]")
+    general_group.add_argument("--disable-protection", dest="disable_protection", action="store_true", default=False,
                 help="Specify this flag to disable the use protection patterns.")
-    general.add_option("--merge-glycine", action="store_true", dest="merge_glycine", default=False,
+    general_group.add_argument("--merge-glycine", action="store_true", dest="merge_glycine", default=False,
                 help="Merge a glycine to the neighbor fragment when fragmenting proteins.")
-    general.add_option("--charge-model", dest="charge_model", default=cfg.getChargeModel(),
-                      help="Charge model to use [%default]")
-    general.add_option("--combine-fragments", dest="combinefragments", type=str, default="",metavar="list of integers",
+    general_group.add_argument("--charge-model", dest="charge_model", default=cfg.getChargeModel(),
+                      help="Charge model to use [%(default)s]")
+    general_group.add_argument("--combine-fragments", dest="combinefragments", type=str, default="",metavar="LIST OF INTEGERS",
                        help="Combines several fragments into one.")
-    output.add_option("--output-format", dest="format", type=str, default=cfg.getWriter(),
-                help="Output format [%default]")
-    output.add_option("--output-boundaries", dest="boundaries", type=str, default="",metavar="list of floats",
+
+    output = parser.add_argument_group("Output Options", description="""
+Specifies which output format (and options of that format) to make when fragit is finished with the fragmentation.
+Some options only apply for some formats.
+The XYZ-MFCC format performes the molecular fractionation with conjugate caps procedure such that the produced fragments
+are capped (when nescessary) and concaps are created (when nescessary).
+""")
+    output.add_argument("--output-format", dest="format", type=str, default=cfg.getWriter(),
+                help="Output format [%(default)s]", choices=supported_output_formats().keys())
+    output.add_argument("--output-boundaries", dest="boundaries", type=str, default="",metavar="list of floats",
                 help="Specifies boundaries for multiple layers. Must be used with --central-fragment option")
-    output.add_option("--output-central-fragment", dest="central_fragment", type=int, default=cfg.getCentralFragmentID(), metavar="integer",
+    output.add_argument("--output-central-fragment", dest="central_fragment", type=int, default=cfg.getCentralFragmentID(), metavar="integer",
                 help="Specifies the fragment to use as the central one. Used in combination with --output-boundaries to make layered inputs")
-    output.add_option("--output-active-distance", dest="active_atoms_distance", type=float, default=cfg.getActiveAtomsDistance(), metavar="float",
-                help="Atoms within this distance from --output-central-fragment will be active. Use with --output-buffer-distance to add buffer region between active and frozen parts. [default: %default]")
-    output.add_option("--output-buffer-distance", dest="maximum_buffer_distance", type=float, default=cfg.getBufferDistance(), metavar="float",
-                help="Maximum distance in angstrom from active fragments from which to include nearby fragments as buffers. This option adds and extends to --output-boundaries. [default: %default]")
-    output.add_option("--output-freeze-backbone", dest="freeze_backbone", action="store_true", default=cfg.getFreezeBackbone(),
+    output.add_argument("--output-active-distance", dest="active_atoms_distance", type=float, default=cfg.getActiveAtomsDistance(), metavar="float",
+                help="Atoms within this distance from --output-central-fragment will be active. Use with --output-buffer-distance to add buffer region between active and frozen parts. [default: %(default)s]")
+    output.add_argument("--output-buffer-distance", dest="maximum_buffer_distance", type=float, default=cfg.getBufferDistance(), metavar="float",
+                help="Maximum distance in angstrom from active fragments from which to include nearby fragments as buffers. This option adds and extends to --output-boundaries. [default: %(default)s]")
+    output.add_argument("--output-freeze-backbone", dest="freeze_backbone", action="store_true", default=cfg.getFreezeBackbone(),
                 help="Option to freeze the backbone of the active region.")
-    output.add_option("--output-jmol-script", dest="output_jmol_script", action="store_true", default=cfg.getWriteJmolScript(),
+    output.add_argument("--output-jmol-script", dest="output_jmol_script", action="store_true", default=cfg.getWriteJmolScript(),
                 help="Write a complimentary jmol script for visualization.")
-    output.add_option("--output-pymol-script", dest="output_pymol_script", action="store_true", default=cfg.getWritePymolScript(),
+    output.add_argument("--output-pymol-script", dest="output_pymol_script", action="store_true", default=cfg.getWritePymolScript(),
                 help="Write a complimentary pymol script for visualization.")
 
-    parser.add_option_group(configuration)
-    parser.add_option_group(general)
-    parser.add_option_group(output)
-    (options, args) = parser.parse_args(argv)
+    args = parser.parse_args()
+    print args
 
-    if len(args) == 0 and len(options.makeconfigfile) > 0:
-        cfg.writeConfigurationToFile(options.makeconfigfile)
-        sys.exit()
-
-    if len(args) != 1:
-        parser.print_help()
-        sys.exit()
-
-    infile = args[0]
+    infile = args.input
 
     molecule = fileToMol(infile)
     conffile = None
-    if len(options.useconfigfile) > 0:
-        conffile = options.useconfigfile
+    if len(args.useconfigfile) > 0:
+        conffile = args.useconfigfile
     fragmentation = Fragmentation(molecule, conffile=conffile)
 
 
     # if there is a config file, read it and ignore other command line options
-    if len(options.useconfigfile) > 0:
+    if len(args.useconfigfile) > 0:
         (writer, output_extension) = get_writer_and_extension(fragmentation.getOutputFormat())
     else:
-        fragmentation.setChargeModel(options.charge_model)
-        fragmentation.setMaximumFragmentSize(options.maxFragmentSize)
-        fragmentation.setOutputFormat(options.format)
-        if options.groupcount > 1: fragmentation.setFragmentGroupCount(options.groupcount)
+        fragmentation.setChargeModel(args.charge_model)
+        fragmentation.setMaximumFragmentSize(args.maxFragmentSize)
+        fragmentation.setOutputFormat(args.format)
+        if args.groupcount > 1: fragmentation.setFragmentGroupCount(args.groupcount)
 
-        (writer, output_extension) = get_writer_and_extension(options.format)
+        (writer, output_extension) = get_writer_and_extension(args.format)
 
     outfile = "%s%s" % (file_basename(infile), output_extension)
-    if len(options.outputfile) > 0:
-        outfile = options.outputfile
+    if len(args.outputfile) > 0:
+        outfile = args.outputfile
 
 
     # do the fragmentation procedure
-    fragmentation.setCombineFragments(options.combinefragments)
-    if options.disable_protection:
+    fragmentation.setCombineFragments(args.combinefragments)
+    if args.disable_protection:
         fragmentation.clearProtectPatterns()
-    if options.merge_glycine:
+    if args.merge_glycine:
         fragmentation.enableMergeGlycinePattern()
     fragmentation.beginFragmentation()
     fragmentation.doFragmentation()
@@ -139,16 +123,16 @@ def main(argv=None):
     out = writer(fragmentation)
 
     # set options from command line
-    boundaries = options.boundaries
-    central_fragment = options.central_fragment
-    active_atoms_distance = options.active_atoms_distance
-    maximum_buffer_distance = options.maximum_buffer_distance
-    freeze_backbone = options.freeze_backbone
-    output_pymol_script = options.output_pymol_script
-    output_jmol_script = options.output_jmol_script
+    boundaries = args.boundaries
+    central_fragment = args.central_fragment
+    active_atoms_distance = args.active_atoms_distance
+    maximum_buffer_distance = args.maximum_buffer_distance
+    freeze_backbone = args.freeze_backbone
+    output_pymol_script = args.output_pymol_script
+    output_jmol_script = args.output_jmol_script
 
     # set options from config file
-    if len(options.useconfigfile) > 0:
+    if len(args.useconfigfile) > 0:
         boundaries = fragmentation.getBoundaries()
         central_fragment = fragmentation.getCentralFragmentID()
         output_pymol_script = fragmentation.getWritePymolScript()
@@ -170,6 +154,6 @@ def main(argv=None):
     out.writeFile(outfile)
 
     # write configuration file
-    if len(options.makeconfigfile) > 0:
-        fragmentation.setBoundaries(boundaries)
-        fragmentation.writeConfigurationToFile(options.makeconfigfile)
+    #if len(args.makeconfigfile) > 0:
+    #    fragmentation.setBoundaries(boundaries)
+    #    fragmentation.writeConfigurationToFile(args.makeconfigfile)